+++ backend/app/models/validacaoParecer/ParecerValidadoGeoGerente.java
@@ -1,6 +1,7 @@
 package models.validacaoParecer;
 
 import models.AnaliseGeo;
+import models.ParecerAnalistaGeo;
 import models.TipoResultadoAnalise;
 import models.UsuarioAnalise;
 import models.tramitacao.AcaoTramitacao;
@@ -15,22 +16,20 @@ public class ParecerValidadoGeoGerente extends TipoResultadoAnaliseChain<Analise
     @Override
     protected void validaParecer(AnaliseGeo analiseGeo, AnaliseGeo novaAnaliseGeo, UsuarioAnalise usuarioExecutor) {
 
-        analiseGeo.tipoResultadoValidacaoGerente = novaAnaliseGeo.tipoResultadoValidacaoGerente;
-        analiseGeo.parecerValidacaoGerente = novaAnaliseGeo.parecerValidacaoGerente;
-        analiseGeo.usuarioValidacaoGerente = usuarioExecutor;
-
         analiseGeo.validarTipoResultadoValidacaoGerente();
 
         analiseGeo._save();
 
-        if (analiseGeo.tipoResultadoAnalise.id == TipoResultadoAnalise.INDEFERIDO) {
+        ParecerAnalistaGeo parecerAnalistaGeo = ParecerAnalistaGeo.find("analiseGeo", analiseGeo).first();
+
+        if (parecerAnalistaGeo.tipoResultadoAnalise.id == TipoResultadoAnalise.INDEFERIDO) {
 
             analiseGeo.analise.processo.tramitacao.tramitar(analiseGeo.analise.processo, AcaoTramitacao.VALIDAR_INDEFERIMENTO_GEO_PELO_GERENTE, usuarioExecutor);
             HistoricoTramitacao.setSetor(HistoricoTramitacao.getUltimaTramitacao(analiseGeo.analise.processo.objetoTramitavel.id), usuarioExecutor);
             return;
         }
 
-        if (analiseGeo.tipoResultadoAnalise.id == TipoResultadoAnalise.DEFERIDO) {
+        if (parecerAnalistaGeo.tipoResultadoAnalise.id == TipoResultadoAnalise.DEFERIDO) {
 
             analiseGeo.analise.processo.tramitacao.tramitar(analiseGeo.analise.processo, AcaoTramitacao.VALIDAR_DEFERIMENTO_GEO_PELO_GERENTE, usuarioExecutor);
             HistoricoTramitacao.setSetor(HistoricoTramitacao.getUltimaTramitacao(analiseGeo.analise.processo.objetoTramitavel.id), usuarioExecutor);