+++ frontend/src/services/validacaoAnaliseGerente.service.js
@@ -3,7 +3,7 @@ var ValidacaoAnaliseGerenteService = function(request,config) {
 	this.concluir = function(params){
 
 		return request
-			.post(config.BASE_URL() + 'analisesGeo/gerente/concluirParecerGerente', params);
+			.post(config.BASE_URL() + 'parecer/gerente/concluirParecerGerente', params);
 	};
 
 };